int to byte

+++ RGBW_controller_and_thermometer/RGBW_controller_and_thermometer.ino
@@ -69,16 +69,16 @@ todo:
 #endif
 
 //Internal Variables
-int ledState = 0; //0 for turning off, 1 for turning on
-int ledUpdate = 1;
+byte ledState = 0; //0 for turning off, 1 for turning on
+byte ledUpdate = 1;
 float ledP = 0; //Led Intensity 1-255 Don't adjust
-int screenPage = 1; //What page to be displayed on the screen
-int ledC[5] = {255}; //Fill with a default value
+byte screenPage = 1; //What page to be displayed on the screen
+byte ledC[5] = {255}; //Fill with a default value
 int ledTarget[5] = {0};
-int oldTimer = 100;
+byte oldTimer = 100;
 int convOnTimes[times]; //Stores the timer on times in a function compatible format
 int convOffTimes[times]; //Store the timer off times in a function compatible format
-int currentTimer = 0; //The current timer being used. Determines what colors should be used.
+byte currentTimer = 0; //The current timer being used. Determines what colors should be used.
 bool oldState = false; //Just used for triggering
 bool firstRun = true; //Label for first loop
 bool ledHold = false; //Hold led adjustment
